+++ src/ui/CurrentSongNode.cpp
@@ -54,6 +54,8 @@ bool CurrentSongNode::init() {
     m_hintSprite->setScale(.8f);
     m_mainLayer->addChildAtPosition(m_hintSprite, geode::Anchor::BottomRight, { 3.f, -6.f });
 
+    m_currentSongLabel = nullptr;
+
     update(0.f);
     scheduleUpdate();
 
@@ -67,13 +69,23 @@ void CurrentSongNode::onButton(cocos2d::CCObject* sender) {
 }
 
 void CurrentSongNode::update(float dt) {
-    auto currentSong = AudioManager::get().getCurrentSong();
-    if (m_currentSong == currentSong) return;
-    m_currentSong = currentSong;
+    auto am = AudioManager::get();
+
+    std::string currentSongLabelContents;
+
+    if (!am.shouldAllowAudioFunctions()) {
+        if (m_currentSongLabel) return; // this has already gone through the first frame, dont do anything we dont need to
+        currentSongLabelContents = "No songs loaded!";
+    } else {
+        auto currentSong = am.getCurrentSong();
+        if (m_currentSong == currentSong) return;
+        m_currentSong = currentSong;
+        currentSongLabelContents = m_currentSong->m_name;
+    }
 
     if (m_currentSongButton) m_currentSongButton->removeFromParent();
 
-    m_currentSongLabel = cocos2d::CCLabelBMFont::create(m_currentSong->m_name.c_str(), "bigFont.fnt");
+    m_currentSongLabel = cocos2d::CCLabelBMFont::create(currentSongLabelContents.c_str(), "bigFont.fnt");
     m_currentSongLabel->limitLabelWidth(178.f, .625f, 0.f);
     m_currentSongLabel->setID("current-song-label");